+++ libs/dh/admin/feature-user-management/src/lib/invite/invite.component.ts
@@ -35,7 +35,6 @@ import { translate, TranslocoDirective, TranslocoService } from '@ngneat/translo
 import { WATT_STEPPER } from '@energinet-datahub/watt/stepper';
 import { WattToastService } from '@energinet-datahub/watt/toast';
 import { WattIconComponent } from '@energinet-datahub/watt/icon';
-import { WattButtonComponent } from '@energinet-datahub/watt/button';
 import { WattFieldErrorComponent } from '@energinet-datahub/watt/field';
 import { WattDropdownComponent, WattDropdownOptions } from '@energinet-datahub/watt/dropdown';
 import { WattTextFieldComponent } from '@energinet-datahub/watt/text-field';
@@ -60,7 +59,7 @@ import {
   readApiErrorResponse,
 } from '@energinet-datahub/dh/market-participant/data-access-api';
 
-import { DhAssignableUserRolesComponent } from './dh-assignable-user-roles/dh-assignable-user-roles.component';
+import { DhAssignableUserRolesComponent } from './assignable-user-roles/assignable-user-roles.component';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -76,7 +75,6 @@ import { DhAssignableUserRolesComponent } from './dh-assignable-user-roles/dh-as
     WATT_MODAL,
     WATT_STEPPER,
     WattIconComponent,
-    WattButtonComponent,
     WattDropdownComponent,
     WattTextFieldComponent,
     WattFieldErrorComponent,
