--- libs/dh/admin/feature-user-management/src/lib/overview/dh-users-overview.component.html
@@ -1,73 +1,0 @@
-<!--
-@license
-Copyright 2020 Energinet DataHub A/S
-
-Licensed under the Apache License, Version 2.0 (the "License2");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<watt-card *transloco="let t; read: 'admin.userManagement.tabs.users'" vater inset="ml">
-  <vater-flex fill="vertical" gap="m">
-    <vater-stack direction="row" gap="s">
-      <h3>{{ t("tabLabel") }}</h3>
-      <span class="watt-chip-label">{{ totalUserCount$ | push }}</span>
-
-      <vater-spacer />
-
-      <watt-search [label]="'shared.search' | transloco" (search)="searchInput$.next($event)" />
-
-      <watt-button
-        *dhPermissionRequired="['fas']"
-        icon="download"
-        variant="text"
-        (click)="download()"
-        [loading]="isDownloading()"
-        >{{ "shared.download" | transloco }}</watt-button
-      >
-
-      <watt-button
-        *dhPermissionRequired="['users:manage']"
-        icon="plus"
-        variant="secondary"
-        [title]="t('inviteUser')"
-        (click)="showInviteUserModal()"
-        >{{ t("inviteUser") }}</watt-button
-      >
-    </vater-stack>
-
-    <vater-stack direction="row" gap="m">
-      <dh-users-overview-filters
-        [statusValue]="initialStatusValue$ | push"
-        [actorOptions]="actorOptions()"
-        [userRoleOptions]="userRolesOptions$ | push"
-        [canChooseMultipleActors]="canChooseMultipleActors()"
-        (filtersChanges)="updateFilters($event)"
-      />
-    </vater-stack>
-
-    <ng-container *rxLet="users$ as users">
-      <dh-users-overview-table
-        [users]="users"
-        [sortChanged]="sortChanged"
-        [isLoading]="isLoading$ | push"
-        [hasGeneralError]="hasGeneralError$ | push"
-        (reload)="reloadUsers()"
-      />
-    </ng-container>
-
-    <watt-paginator
-      [length]="totalUserCount$ | push"
-      [pageSize]="pageSize$ | push"
-      [pageIndex]="pageIndex$ | push"
-      (changed)="onPageChange($event)"
-    />
-  </vater-flex>
-</watt-card>
