+++ libs/dh/admin/feature-user-management/src/lib/table/users-table.component.ts
@@ -53,12 +53,12 @@ import {
 import { DhEmDashFallbackPipe } from '@energinet-datahub/dh/shared/ui-util';
 
 import { DhUserDrawerComponent } from '../details/details.component';
-import { DhUserLatestLoginComponent } from './dh-user-latest-login.component';
+import { DhUserLatestLoginComponent } from '../dh-user-latest-login.component';
 
 @Component({
   selector: 'dh-users-overview-table',
   standalone: true,
-  templateUrl: './dh-users-overview-table.component.html',
+  templateUrl: './users-table.component.html',
   styles: [
     `
       :host {
